refactor: split config module into package
