+++ baseball_data_lab/apis/local_data_client.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 
-from baseball_data_lab.config import DATA_DIR
+from baseball_data_lab.config.paths import DATA_DIR
 
 class LocalDataClient:
