+++ baseball_data_lab/data_converter.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 
-from baseball_data_lab.config import DATA_DIR
+from baseball_data_lab.config.paths import DATA_DIR
 
 
 class DataConverter:
